LangGraph_1o1_Agentic_Customer_Support/requirements.txt
gradio==4.41.0
ipykernel==6.29.5
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langgraph==0.2.4
langgraph-checkpoint==1.0.3
langsmith==0.1.99
numpy==1.26.4
openai==1.41.0
pandas==2.2.2
pillow==10.4.0
pyprojroot==0.3.0
PyYAML==6.0.2
SQLAlchemy==2.0.32

